Demo apps

+++ pkg/controllers/system_jobs.go
@@ -51,8 +51,10 @@ func (job SystemJob) Run() {
 			continue
 		}
 		compIDs := difference(dbComps, comps)
-		if err := db.DeleteComponentsWithIDs(compIDs, time.Now()); err != nil {
-			logger.Errorf("error deleting components: %v", err)
+		if compIDs != nil {
+			if err := db.DeleteComponentsWithIDs(compIDs, time.Now()); err != nil {
+				logger.Errorf("error deleting components: %v", err)
+			}
 		}
 	}
 	topology.ComponentRun()
